requirements.txt
# Core dependencies
twitchio==2.6.0
python-dotenv==1.0.0
openai==1.10.0
motor==3.3.2
websockets==12.0
aiohttp==3.9.3
pydantic==2.5.3
twitchAPI>=4.0.0
fastapi
uvicorn[standard]


# Utility packages
cachetools==5.3.2
backoff==2.2.1
async-timeout==4.0.3
PyYAML==6.0.1

# Error tracking
sentry-sdk==1.39.1

# Type hints support
typing-extensions==4.9.0

# Logging enhancements
colorlog==6.8.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.0
pytest-cov==4.1.0

# Development tools (optional)
black==24.1.1
flake8==7.0.0
isort==5.13.2


